[Lab 08a] 06 - Refactored Package Structure, moved MVP Files from .activities to .views.<subfolder>
