+++ app/src/main/res/layout/activity_spot.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.archaeologicalfieldwork.activities.SpotView">
+    tools:context="com.archaeologicalfieldwork.views.spot.SpotView">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/appBarLayout"
