+++ app/src/main/res/layout/activity_spot_list.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.archaeologicalfieldwork.activities.SpotListView">
+    tools:context="com.archaeologicalfieldwork.views.spotList.SpotListView">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
